firmware: add experimental software debounce.
Suspected EMI transients have been observed spuriously triggering the
external interrupt lines on some systems.  Likely this is better solved
in hardware, however this is not as convenient as a software fix.

When enabled, ISRs first delay by a fixed about (~1.2ms currently) and
then inspects the active-low interrupting pin.  The delay amount
was selected somewhat empirically (with a non-SF800 sensor) and
likely requires tuning to account for differences in sensors.